feat: realizando a requisicao GET para encontrar todos os Users em "/users", por meio da logica de servico da Classe UserService, que foi implementada para fazer a validacao dos dados e chamar o Repository

+++ course/src/main/java/com/joao/neto/ufcg/course/entities/User.java
@@ -13,7 +13,7 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
     private String phone;
@@ -21,7 +21,7 @@ public class User implements Serializable {
 
     public User() {}
 
-    public User(Integer id, String name, String email, String phone, String password) {
+    public User(Long id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -52,7 +52,7 @@ public class User implements Serializable {
                 '}';
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
